+++ src/middlewares/index.ts
@@ -73,5 +73,6 @@ export const defaultValidationFailedHandler: ValidationFailedHandler = async (er
 export * from './buffer';
 export * from './cookies';
 export * from './json';
+export * from './lang';
 export * from './query';
 export * from './validate';